Close #14: Make release workflow also copy nix result directory
Made the publish release workflow also copy the nix result directory into
the artifact directory. This allows for the ability of the artifact publisher
to search for files in the result directory.

A very clear use case is: let's say that the build produces a PDF. Now when
you create a release you want the PDF to be a artifact that can be downloaded,
not a tarball that has to be decompressed and then searched. That is not good
UX. This solves that problem as the workflow can now publish any file within
the artifact directory that matches the provided globs.