fix: failing workflow by removed artifact paths
The paths provided in '.github/workflows/publish-release.yml' are all required
instead of optional. I thought that if you add a path there It'd add that to
the paths it'd search in for artifacts.

However instead of failing if there are no files at all, it fails whenever any
of these search paths are empty, which is not what I want.