Added an update-flake-lock workflow
I added this workflow to keep the flake.lock file up to date as 'flake-checker'
seemed to have an issue with the flake.lock file being out of date. So now I
can't push anymore on those branches without ignoring the commit hooks all
together which defeats the entire purpose.