fix: flake lock update PR automation
It was failing weekly. Now it is copied directly from the their repository.
So if it still does not work, I don't know why. It uses a unique branch so
there should never be any collisions. The only issue I can see with it, is
that the PR will not trigger any checks, and thus will not perform a
`nix flake check`. Though the odds of this breaking something is small, it
is still not something you want to happen. I am looking into how to fix this
without any user input at all. But the only options seem to either add a PAT
(Personal Access Token), or open/reclose the PR to trigger the workflows.