ci: fixed broken sync label workflow
The workflow was failing because of a mismatch between the version/clis on my
machine and the workflow itself. Causing the workflow to read invalid JSON,
namely YAML as their version of `yq` outputs YAML by default, and mine JSON.
Secondly, their version of `gh` does not support the `-y` flag, and mine does.